% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{collections}
\alias{collections}
\alias{collections_default_columns}
\alias{collections_detail}
\alias{collection_contributors}
\alias{collection_datasets}
\title{HuBMAP Collections}
\usage{
collections()

collections_default_columns(as = c("character", "tibble"))

collections_detail(uuid)

collection_contributors(uuid)

collection_datasets(uuid)
}
\arguments{
\item{as}{character(1) return format. One of \code{"tibble"} (default),
or \code{"character"}.}

\item{uuid}{character(1) corresponding to the HuBMAP UUID
string. This is expected to be a 32-digit hex number.}
}
\value{
\verb{*_columns()} returns a named list \code{name}
containing the column name used in the tibble returned by
\code{samples()}, \code{datasets()},  \code{donors()}, \code{collections()},
or \code{publications()}. When \code{as = "tibble"}, the return value is a tibble
with paths as elements and abbreviations as names.
}
\description{
\code{collections()} returns details about available collections

\verb{*_columns()} returns a tibble or named
character vector describing the content of the tibble returned
by \code{samples()}, \code{datasets()},  \code{donors()}, \code{collections()},
or \code{publications()}.

\code{collections_detail()} takes a unique collection_id and
returns details about one specified collection as a tibble

See \code{collection_information()} and \code{collection_title()} to
easily summarize a collection from its project id.

\code{collection_contributors()} takes a unique collection_id and
returns contributors information of one specified collection as a tibble

\code{collection_datasets()} takes a unique collection_id and
returns contributors information of one specified collection as a tibble
}
\details{
Additional details are provided on the HuBMAP consortium
webpage, https://software.docs.hubmapconsortium.org/apis

Additional details are provided on the HuBMAP consortium
webpage, https://software.docs.hubmapconsortium.org/apis

Additional details are provided on the HuBMAP consortium
webpage, https://software.docs.hubmapconsortium.org/apis

Additional details are provided on the HuBMAP consortium
webpage, https://software.docs.hubmapconsortium.org/apis
}
\examples{
collections()
collections_default_columns()

uuid <-
    collections() |>
    dplyr::slice(1) |>
    dplyr::pull("uuid")
    
collections_detail(uuid)
uuid <-
    collections() |>
    dplyr::slice(1) |>
    dplyr::pull("uuid")
    
collection_contributors(uuid)
uuid <-
    collections() |>
    dplyr::slice(1) |>
    dplyr::pull("uuid")
    
collection_datasets(uuid)
}
