% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{bulk_data_transfer}
\alias{bulk_data_transfer}
\title{HuBMAP File Querying}
\usage{
bulk_data_transfer(uuid)
}
\arguments{
\item{uuid}{character(1) corresponding to the HuBMAP Dataset UUID
string. This is expected to be a 32-digit hex number. One unique Dataset
UUID corresponds to one specific URL on Globus. Files corresponding to
unique Dataset UUID are saved on Globus HuBMAP Public Collection webpage.}
}
\value{
\code{bulk_data_transfer()} display information about how to bulk
download all files related to the dataset or publication.
}
\description{
\code{bulk_data_transfer()} takes a unique dataset or publication
UUID to be used to query the HuBMAP API for information about available
files.
}
\details{
Additional details are provided on the HuBMAP consortium
webpage, https://software.docs.hubmapconsortium.org/apis
}
\examples{
# Globus available
uuid_globus <- "2d3dbd5e5a3d9d5cfc7a46f06815a4eb"
uuid_sra_dbGaP <- "5ba6a5b81e95c93c26a33980f6e957d7"
uuid_not_available <- "0eb5e457b4855ce28531bc97147196b6"

# bulk_data_transfer(*_globus)

}
