% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{datasets}
\alias{datasets}
\alias{samples}
\alias{donors}
\title{HuBMAP dataset, sample, and donor metadata}
\usage{
datasets(columns = c("portal", "some", "all"))

samples(columns = c("portal", "some", "all"))

donors(columns = c("portal", "all"))
}
\arguments{
\item{columns}{character(1) columns to extract from the
database. \code{"portal"} (default): columns used in the
corresponding page of the HuBMAP portal. \code{"some"} (available
for \code{datasets()} and \code{samples()}): columns present in all
entities; \code{"all"}: all columns in the database.}
}
\description{
HuBMAP dataset, sample, and donor metadata
}
\examples{
datasets()
datasets() |>
    dplyr::glimpse()
datasets() |>
    dplyr::count(organ, sort = TRUE)
samples()
samples() |>
    dplyr::glimpse()
samples() |>
    dplyr::count("mapped_specimen_type", sort = TRUE)
donors()
donors() |>
    dplyr::glimpse()

donors() |>
    dplyr::count(race, sort = TRUE)

## number of non-missing values
donors() |>
    dplyr::summarize(dplyr::across(.fns = ~sum(!is.na(.x)))) |>
    dplyr::glimpse()

## mostly complete data
donors(columns = "all") |>
    dplyr::select(where(~ sum(!is.na(.x)) > 100L))
}
