% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organ.R
\name{organ}
\alias{organ}
\title{HuBMAP Organs}
\usage{
organ()
}
\value{
\code{organ} returns the tibble with all available organs' abbreviations
and corresponding full name.
}
\description{
\code{organ} retrieves available organ categories and abbreviations
}
\examples{
organs <- organ() 
organs |>
    dplyr::select("name") |>
    dplyr::distinct()
}
