% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance.R
\name{uuid_provenance}
\alias{uuid_provenance}
\title{HuBMAP Provenance}
\usage{
uuid_provenance(uuid)
}
\arguments{
\item{uuid}{character(1) corresponding to the HuBMAP Sample UUID
string. This is expected to be a 32-digit hex number.}
}
\value{
\code{uuid_provenance()} takes a Dataset, Sample, or Donor UUID,
and returns the provenance details as a list of characters, from the most
recent ancestor to the furthest ancestor. All entity types are included.
}
\description{
HuBMAP Provenance
}
\details{
Additional details are provided on the HuBMAP consortium
webpage, https://software.docs.hubmapconsortium.org/apis
}
\examples{
uuid <- "acf816bed56dd6ba55da82960d937ec4"
uuid_provenance(uuid)
}
