% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publications.R
\name{publications}
\alias{publications}
\alias{publications_default_columns}
\alias{publications_detail}
\alias{publication_contributors}
\title{HuBMAP Samples}
\usage{
publications()

publications_default_columns(as = c("character", "tibble"))

publications_detail(uuid)

publication_contributors(uuid)
}
\arguments{
\item{as}{character(1) return format. One of \code{"tibble"} (default),
or \code{"character"}.}

\item{uuid}{character(1) corresponding to the HuBMAP UUID
string. This is expected to be a 32-digit hex number.}
}
\value{
\verb{*_columns()} returns a named list \code{name}
containing the column name used in the tibble returned by
\code{samples()}, \code{datasets()}, \code{donors()}, \code{collections()}, or
\code{publications()}. When \code{as = "tibble"}, the return value is a tibble
with paths as elements and abbreviations as names.
}
\description{
\code{publications()} returns details about available samples

\verb{*_columns()} returns a tibble or named
character vector describing the content of the tibble returned
by \code{samples()}, \code{datasets()}, \code{donors()}, \code{collections()},
or \code{publications()}.

\code{publications_detail()} takes a unique publication_id and
returns details about one specified publication as a tibble

\code{publication_contributors()} takes a unique collection_id and
returns contributors information of one specified collection as a tibble
}
\details{
Additional details are provided on the HuBMAP consortium
webpage, https://software.docs.hubmapconsortium.org/apis

Additional details are provided on the HuBMAP consortium
webpage, https://software.docs.hubmapconsortium.org/apis

Additional details are provided on the HuBMAP consortium
webpage, https://software.docs.hubmapconsortium.org/apis
}
\examples{
publications()
publications_default_columns()

uuid <-
    publications() |>
    dplyr::slice(1) |>
    dplyr::pull("uuid")
    
publications_detail(uuid)
uuid <-
    publications() |>
    dplyr::slice(1) |>
    dplyr::pull("uuid")
    
publication_contributors(uuid)
}
