% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{samples}
\alias{samples}
\alias{samples_default_columns}
\alias{samples_detail}
\title{HuBMAP Samples}
\usage{
samples()

samples_default_columns(as = c("character", "tibble"))

samples_detail(uuid)
}
\arguments{
\item{as}{character(1) return format. One of \code{"tibble"} (default),
or \code{"character"}.}

\item{uuid}{character(1) corresponding to the HuBMAP UUID
string. This is expected to be a 32-digit hex number.}
}
\value{
\verb{*_columns()} returns a named list \code{name}
containing the column name used in the tibble returned by
\code{samples()}, \code{datasets()},  \code{donors()}, \code{collections()},
or \code{publications()}. When \code{as = "tibble"}, the return value is a tibble
with paths as elements and abbreviations as names.
}
\description{
\code{samples()} returns details about available samples

\verb{*_columns()} returns a tibble or named
character vector describing the content of the tibble returned
by \code{samples()}, \code{datasets()},  \code{donors()}, \code{collections()},
or \code{publications()}.

\code{samples_detail()} takes a unique sample_id and
returns details about one specified sample as a tibble
}
\details{
Additional details are provided on the HuBMAP consortium
webpage, https://software.docs.hubmapconsortium.org/apis

Additional details are provided on the HuBMAP consortium
webpage, https://software.docs.hubmapconsortium.org/apis
}
\examples{
samples()
samples() |>
    dplyr::glimpse()
samples() |>
    dplyr::count("sample_category", sort = TRUE)
samples_default_columns()

uuid <-
    samples() |>
    dplyr::slice(1) |>
    dplyr::pull("uuid")
    
samples_detail(uuid)
}
