% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset}
\alias{dataset}
\title{Retrieve information about a single dataset}
\usage{
dataset(uuid)
}
\description{
Retrieve information about a single dataset
}
\examples{
uuid <-
    datasets() |>
    dplyr::slice(1) |>
    dplyr::pull("uuid")
}
