% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{datasets}
\alias{datasets}
\alias{datasets_default_columns}
\alias{dataset_detail}
\alias{dataset_derived}
\title{HuBMAP Datasets}
\usage{
datasets(size = 10000L, from = 0L)

datasets_default_columns(as = c("tibble", "character"))

dataset_detail(uuid)

dataset_derived(uuid)
}
\arguments{
\item{size}{integer(1) number of maximum results to return;
The default (10000) is meant to be large enough to return all results.}

\item{from}{integer(1) number of number of results to skip, defaulting to 0.}

\item{as}{character(1) return format. One of \code{"tibble"} (default),
or \code{"character"}.}

\item{uuid}{character(1) corresponding to the HuBMAP Sample UUID
string. This is expected to be a 32-digit hex number.}
}
\value{
\verb{*_columns()} returns a named list \code{name}
containing the column name used in the tibble returned by
\code{samples()}, \code{datasets()},  \code{donors()},
\code{collections()},  or \code{publications()}.
When \code{as = "tibble"},the return value is a tibble
with paths as elements and abbreviations as names.
}
\description{
\code{datasets} returns the details available datasets, ordered by
last modified dates

\verb{*_columns()} returns a tibble or named
character vector describing the content of the tibble returned
by \code{samples()}, \code{datasets()},  \code{donors()}, \code{collections()},
or \code{publications()}.

\code{dataset_detail()} takes a unique dataset_id and
returns details about one specified dataset as a tibble

\code{dataset_derived()} takes a unique sample_id and
returns the derived (support) dataset details. Support datasets normally
belong to Image Pyramid, with image files available to download via
Globus Collection. See details to download in \code{files_globus_url()}.
}
\details{
Additional details are provided on the HuBMAP consortium
webpage, https://software.docs.hubmapconsortium.org/apis

Additional details are provided on the HuBMAP consortium
webpage, https://software.docs.hubmapconsortium.org/apis
}
\examples{
datasets(size = 100, from = 200)
datasets_default_columns()

uuid <- "7754aa5ebde628b5e92705e33e74a4ef"
dataset_detail(uuid)
# no derived dataset
uuid <- "3acdb3ed962b2087fbe325514b098101"
dataset_derived(uuid)

# with derived dataset
uuid <- "2c77b1cdf33dbed3dbfb74e4b578300e"
dataset_derived(uuid)
}
