% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{datasets}
\alias{datasets}
\alias{datasets_default_columns}
\alias{datasets_detail}
\title{HuBMAP Datasets}
\usage{
datasets()

datasets_default_columns(as = c("character", "tibble"))

datasets_detail(uuid)
}
\arguments{
\item{as}{character(1) return format. One of \code{"tibble"} (default),
or \code{"character"}.}

\item{uuid}{character(1) corresponding to the HuBMAP UUID
string. This is expected to be a 32-digit hex number.}
}
\value{
\verb{*_columns()} returns a named list with column \code{name}
containing the column name used in the tibble returned by
\code{samples()}, \code{datasets()},  \code{donors()}, \code{collections()},
or \code{publications()}. When \code{as = "tibble"}, the return value is a tibble
with paths as elements and abbreviations as names.
}
\description{
\code{datasets} returns the details available datasets

\verb{*_columns()} returns a tibble or named
character vector describing the content of the tibble returned
by \code{samples()}, \code{datasets()},  \code{donors()}, \code{collections()},
or \code{publications()}.

\code{datasets_detail()} takes a unique dataset_id and
returns details about one specified dataset as a tibble
}
\details{
Additional details are provided on the HuBMAP consortium
webpage, https://software.docs.hubmapconsortium.org/apis
}
\examples{
datasets()
datasets() |>
    dplyr::glimpse()
datasets() |>
    dplyr::count(dataset_type, sort = TRUE)
datasets_default_columns()

uuid <-
    datasets() |>
    dplyr::slice(1) |>
    dplyr::pull("uuid")
    
datasets_detail(uuid)
}
