% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{files_globus_url}
\alias{files_globus_url}
\title{HuBMAP File Querying}
\usage{
files_globus_url(uuid)
}
\arguments{
\item{uuid}{character(1) corresponding to the HuBMAP Dataset UUID
string. This is expected to be a 32-digit hex number. One unique Dataset
UUID corresponds to one specific URL on Globus. Files corresponding to
unique Dataset UUID are saved on Globus HuBMAP Public Collection webpage.}
}
\description{
\code{files_globus_url()} takes a unique dataset UUID or publication UUID
to be used to query theHuBMAP API for information about available files.
}
\details{
Additional details are provided on the HuBMAP consortium
webpage, https://software.docs.hubmapconsortium.org/apis
}
