% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{datasets}
\alias{datasets}
\alias{samples}
\alias{donors}
\title{"HuBMAP dataset, sample, and donor metadata"}
\usage{
datasets(columns = c("portal", "some", "all"))

samples(columns = c("portal", "some", "all"))

donors(columns = c("portal", "all"))
}
\description{
"HuBMAP dataset, sample, and donor metadata"
}
\examples{
datasets()
datasets() |>
    dplyr::glimpse()
# datasets() |>
#      dplyr::count(origin_sample.mapped_organ, sort = TRUE)
samples()
samples() |>
    dplyr::glimpse()
samples() |>
    dplyr::count("mapped_specimen_type", sort = TRUE)
donors()
donors() |>
    dplyr::glimpse()

donors() |>
    dplyr::count(race, sort = TRUE)

## number of non-missing values
donors() |>
    dplyr::summarize(dplyr::across(.fns = ~sum(!is.na(.x)))) |>
    dplyr::glimpse()

## mostly complete data
donors() |>
    dplyr::select(where(~ sum(!is.na(.x)) > 100L))
}
